# Vet clinic database: database performance audit