## Fulfilled Requirements:

- Changed the owner's table by removing the age column and adding the email field.
- Added the provided data to the `owners` and `animals` relation.
- Conducted initial `Explain Analyze` audit on different queries.
- `SELECT COUNT(*) FROM visits where animal_id = 4;`
- `SELECT * FROM visits where vet_id = 2;`
- `SELECT * FROM owners where email = 'owner_18327@mail.com';`
![expl_analyze](https://github.com/Kanyorok/vet-clinicdb/assets/107240977/adb901f3-6909-43b7-a89f-e95d3d53b213)
## Optimization of Performance

- Added index for optimizing query `execution time`.
![index](https://github.com/Kanyorok/vet-clinicdb/assets/107240977/c5d31efb-28a4-41dd-b2d4-20cbfbed4849)
- Time for the first query reduced from 1637.24 ms to 253.692 ms.
![optimize1](https://github.com/Kanyorok/vet-clinicdb/assets/107240977/02002df2-3a7c-4ee0-8d04-ea4e9cbdfa8a)
- Similarly the time for `EXPLAIN ANALYZE SELECT * FROM owners where email = 'owner_18327@mail.com';` reduced from 1163.782 ms to 56.463 ms.
![optimize2](https://github.com/Kanyorok/vet-clinicdb/assets/107240977/cdc45cf6-0446-4623-b452-05d39c2fefd1)

- For the query `SELECT * FROM visits where vet_id = 2;` the execution time increased from 2374.646 ms to 6169.521 ms. We optimized by setting enable_seqscan to `off`. The command `SET enable_seqscan = off;` helped reduce the execution time to    1754.567 ms.
![optimize3](https://github.com/Kanyorok/vet-clinicdb/assets/107240977/783d2b75-4e0e-4e30-bede-03e2eabfd36f)